Better version :)
The CommandType enum was removed because it wasn't being used.
The SlashCommandBuilder was replaced with an object literal to define the command's data. This is simpler and more concise.
The getMemberAndUser method was removed because it wasn't being used. Instead, the target user is retrieved from the interaction's options or default to the invoking user.
The EmbedBuilder was replaced with a plain object to define the embed's options. This is simpler and more concise.
The code that sets the embed's color was moved to a separate getColor function, which is more modular and easier to test. This function tries to get the user's display color from each guild they're in, returning the first non-black color it finds. If it fails, it logs an error and returns undefined.
The client.reply method was replaced with interaction.reply, which is more concise and clearer.